shifted object info cache in front of shortcut checking to halve the average number of read requests to the dynamodb tables, thus lowering response times. also implemented daily periodic purging of the caches at 6am UTC+8. caches might be >1MB which is the limit of table.scan() so tweaked code to repeat scanning until tables are empty. upgraded to discord.py v2.0 (testing version) to make use of tasks.loop() time argument which allows bot to perform tasks daily at specific timings

+++ dynamodb_methods.py
@@ -72,7 +72,13 @@ def ddb_remove_all(table_name):
 	table = ddb_create_session().Table(table_name)
 	key_header = get_table_headers(table_name)[0]
 
-	items = table.scan()['Items']
-	for item in items:
-		table.delete_item(Key={
-			key_header: item[key_header]})
+	# scan has a size limit of 1MB, so repeat until it is empty
+	while True:
+		items = table.scan()['Items']
+
+		if not items:
+			break
+
+		for item in items:
+			table.delete_item(Key={
+				key_header: item[key_header]})
